feat: move `Button` component to shared scope of app module
