+++ src/app/screens/auth/screens/forgot-password/forgot-password-modal.tsx
@@ -4,8 +4,8 @@ import { Formik, Form } from 'formik';
 import * as Yup from 'yup';
 
 import { getFormStatusAndErrors } from 'utils';
-import { AuthApi } from 'app/shared';
-import { Modal, Status, Input, Button } from 'app/screens/auth/shared';
+import { AuthApi, Button } from 'app/shared';
+import { Modal, Status, Input } from 'app/screens/auth/shared';
 
 export type ForgotPasswordModalProps = { onRequestClose?: () => void } & RouteComponentProps<any>;
