+++ src/app/screens/auth/shared/components/index.ts
@@ -1,3 +1,2 @@
 export * from './modal';
 export * from './form';
-export * from './button';
