+++ src/app/shared/components/index.ts
@@ -2,3 +2,4 @@ export * from './chip';
 export * from './backdrop';
 export * from './action-sheet';
 export * from './route';
+export * from './button';
